+++ baelfire/task/tests/test_screen.py
@@ -62,7 +62,7 @@ class TestAttachScreenTask(ScreenFixtures):
 
     @yield_fixture
     def mrun_before(self):
-        with patch('baelfire.task.screen.RunTask') as mock:
+        with patch.object(ExampleAttachScreenTask, 'run_before') as mock:
             yield mock
 
     def test_no_detached_task_configured(self):